+++ ajax.php
@@ -9,7 +9,7 @@ if (!$_SESSION['login_string']) {
 session_write_close();
 
 $CACHED = true;
-if (isset($_GET['cached']) && $_GET['cached'] == 0) {
+if (isset($_GET['cached']) && ($_GET['cached'] == 'false' || $_GET['cached'] == 0)) {
     $CACHED = false;
 }
 $GLOBALS['cached'] = $CACHED;
@@ -17,7 +17,7 @@ $GLOBALS['cached'] = $CACHED;
 $type = (!empty($_GET['type']) ? $_GET['type'] : $_POST['type']);
 $action = (!empty($_GET['action']) ? $_GET['action'] : $_POST['action']);
 
-$type = ucwords(strtolower($type));
+$type = str_replace(' ', '', lcfirst(ucwords(str_replace('-', ' ', strtolower($type)))));
 $action = str_replace(' ', '', lcfirst(ucwords(str_replace('-', ' ', strtolower($action)))));
 
 // If we're not posting, it's a get function