+++ includes/classes/pools/eligius.php
@@ -42,6 +42,11 @@ class Pools_Eligius extends Pools_Abstract {
                 $poolData[$action] = curlCall($this->_apiURL  . '/api.php?cmd='.$actionParam);
             }
 
+            // Offline Check
+            if (empty($poolData[$this->_actions[0]])) {
+                return;
+            }
+
             // Data Order
             $data['type'] = $this->_type;
 
@@ -62,7 +67,7 @@ class Pools_Eligius extends Pools_Abstract {
             $data['last_block'] = $poolData['livedata']['lastblockheight'];
             $data['last_block_url'] = $this->_apiURL  . '/blocks.php';
 
-            $data['url_name'] = $this->_apiURL;
+            $data['url_name'] = 'http://eligius.st';
             $data['url'] = $this->_apiURL . '/userstats.php/' . $this->_btcaddess;
 
             $this->_fileHandler->write(json_encode($data));