+++ js/dashboard/RigCollection.js
@@ -37,22 +37,20 @@
     })
 
     /*==========  Initial data call  ==========*/
-    this._getData(function (data) {
-    //   _self._buildOverview(data)
+    $.when(_self._getData()).done(function (data) {
+        _self._buildOverview(data)
 
-      _self.apiData = { type: 'rigs', action: 'update' }
+        _self.apiData = { type: 'rigs', action: 'update' }
 
-      _self._update() // Populate all devices in the collection
+        _self._update() // Populate all devices in the collection
 
-      /*==========  Setup polling  ==========*/
-      setInterval(function () {
-        if (_self._ready) {
-          _self._rigsResponded = 0
-          _self._ready = false
-          _self._update() // Update both devices and summary stats
-        }
-      }, root.rigUpdateTime)
-    })
+        /*==========  Setup polling  ==========*/
+        setInterval(function () {
+          if (_self._ready) {
+              _self._update() // Update both devices and summary stats
+          }
+      }, root.rigUpdateTime);
+    });
   }
 
   /*-----  End of RigCollection Public Methods  ------*/
@@ -63,42 +61,54 @@
   =====================================================*/
 
   RigCollection.prototype._add = function (rigId) {
-    this.collection.push(new this.SubClass(rigId))
+      this.collection.push(new this.SubClass(rigId))
   }
 
   RigCollection.prototype._update = function () {
     var _self = this // caching self ref for passing down in scope
-    var overviewData = []
+    var overviewData = [];
 
-    this.collection.forEach(function (rig, index) {
-      _self.apiData.id = rig.rigID
-      _self._getData(function (data) {
-        rig.update(data)
-        overviewData[index] = data
-        _self._rigsResponded++
-        if (_self._rigsResponded === _self.collection.length) {
-          _self._ready = true
-          _self._buildOverview(overviewData)
-          $(document).trigger('masonry-update')
-        }
-      })
-    })
+    if (_self._ready) {
+        _self._ready = false;
+        _self._rigsResponded = 0;
 
-    return true;
+        var promises = $.map(this.collection, function (rig, index) {
+            _self.apiData.id = rig.rigID;
+            return _self._getData();
+        });
+
+        $.when.apply($, promises).done(function () {
+            var arrData = arguments;
+            if (arguments.length != _self.collection.length && _self.collection.length == 1) {
+                arrData = [arguments];
+            }
+            overviewData = $.map(arrData, function (arr, idx) {
+                _self.collection[idx].update(arr[0])
+                _self._rigsResponded++;
+                return arr[0];
+            })
+
+            if (_self._rigsResponded === _self.collection.length) {
+                _self._buildOverview(overviewData);
+                $(document).trigger('masonry-update');
+                _self._ready = true;
+            }
+        })
+    } else {
+        return false;
+    }
   }
 
-  RigCollection.prototype._getData = function (callback) {
-    var _self = this // caching self ref for passing down in scope
+  RigCollection.prototype._getData = function () {
+    var _self = this;
 
-    $.ajax({
-      url: 'ajax.php',
-      data : _self.apiData,
-      dataType: 'json'
-    })
-    .done(callback)
-    .fail(function (xhr, status, message) {
-      //console.error(message)
-    })
+    var ajaxPromise = $.ajax({
+        url: 'ajax.php',
+        data: _self.apiData,
+        dataType: 'json'
+    });
+
+    return ajaxPromise;
   }
 
   RigCollection.prototype._buildOverview = function (data) {
@@ -135,8 +145,14 @@
     this.$overviewTableBody.html(this.overviewTableData)
 
     if (cG.showTotalHashrate == 'true') {
+        var algorithmContainer = $('#total-hashrates li a');
+        // get each hashrate
+        var existingAlgorithms = [];
+        algorithmContainer.each(function(k, v) {
+            existingAlgorithms[$(v).attr('id').replace('hashrate_', '')] = k;
+        });
+
         for (var key in algorithms) {
-          if (algorithms[key]) {
             if ($('#hashrate_' + key).length == 0) {
               var aTag = document.createElement('a');
               aTag.setAttribute('id',"hashrate_"+key);
@@ -144,10 +160,15 @@
               $('#total-hashrates li').append(aTag)
             }
             $('#hashrate_' + key).removeClass('hidden').html('<span class="hashrate-algo">' + key + '</span>' + Util.getSpeed(algorithms[key]))
-          }
-          else {
-            $('#total-hashrates #hashrate_' + key).addClass('hidden')
-          }
+
+            if (typeof existingAlgorithms[key] != 'undefined') {
+                delete existingAlgorithms[key];
+            }
+        }
+
+        for (var key in existingAlgorithms) {
+            $('#hashrate_' + key).remove();
+            delete existingAlgorithms[key];
         }
     }