+++ permissionfix.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
-# cryptoGlance permission fix script for apache2
+# cryptoGlance permission fix script for most linux setups
+# http://cryptoglance.info
 
-sudo chown -R www-data *
+sudo chown -R www-data:www-data .
 sudo chmod -R 777 user_data
-sudo service apache2 restart