+++ update.php
@@ -24,6 +24,21 @@ if (isset($_POST['cryptoglance_version']) &&
     // get settings for update type to get
     $settings = $cryptoGlance->getSettings();
 
+    $osInfo = strtolower(php_uname());
+    $osType = '';
+    $phpUser = null;
+    if (stripos($osInfo, 'windows') !== false) {
+        $osType = 'windows';
+    } else if (stripos($osInfo, 'linux') !== false) {
+        $osType = 'linux';
+        if (function_exists('posix_geteuid') && function_exists('posix_getpwuid')) {
+            $phpUserInfo = posix_getpwuid(posix_geteuid());
+            $phpUser = $phpUserInfo['name'];
+        } else {
+            $phpUser = 'www-data';
+        }
+    }
+
     // Need this late in the script
     function lensort($a,$b){
         return strlen($a)-strlen($b);
@@ -59,6 +74,7 @@ if (isset($_POST['cryptoglance_version']) &&
         <pre style=\"font-family: Menlo,Monaco,Consolas,Courier New,monospace;margin:0;white-space: pre-wrap;\">";
     if (ob_get_level() == 0) ob_start();
     if ($settings['general']['updates']['enabled'] == 1) {
+
         $updateType = $settings['general']['updates']['type'];
 
         echo '==> Starting Update...Please be patient<br />'; ob_flush(); flush(); sleep(1);
@@ -67,11 +83,28 @@ if (isset($_POST['cryptoglance_version']) &&
 
         echo '==> Creating temporary directory to download update and unzip: '.$updateDir.'<br />'; ob_flush(); flush();
 
-        if (!mkdir($currentDir . $updateDir, 0777, true)) {
+        if (!mkdir($currentDir . $updateDir, 0777, true) && !file_exists($currentDir . $updateDir)) {
             echo '==> ERROR: Failed to create the directory: ' . $updateDir . '<br />';  ob_flush(); flush(); sleep(1);
-            die('* Please contact support via reddit, bitcointalk, or IRC!');
+            echo '********************' . '<br />';
+            echo 'There is a problem with the user permissions that will not allow CryptoGlance to update.' . '<br />';
+            if ($osType == 'linux') {
+                echo 'Please run the command "sudo sh permissionfix.sh" in the root of the CryptoGlance directory to fix this issue.' . '<br />';
+            }
+            echo '********************' . '<br />';
+            die('* Please contact support via reddit, bitcointalk, or IRC if you need assistance!');
         } else {
             echo '==> Successfully created temporary directory!<br />'; ob_flush(); flush(); sleep(1);
+        }
+
+        // Permission Check
+        if (!is_writable($currentDir . 'includes' . DIRECTORY_SEPARATOR . 'inc.php')) {
+            echo '********************' . '<br />';
+            echo 'There is a problem with the user permissions that will not allow CryptoGlance to update.' . '<br />';
+            if ($osType == 'linux') {
+                echo 'Please run the command "sudo sh permissionfix.sh" in the root of the CryptoGlance directory to fix this issue.' . '<br />';
+            }
+            echo '********************' . '<br />';
+            die('* Please contact support via reddit, bitcointalk, or IRC if you need assistance!');
         }
 
         // DOWNLOADING -----------------------
@@ -157,7 +190,13 @@ if (isset($_POST['cryptoglance_version']) &&
                     echo '==> Deleted File: ' . $realFilePath . '<br />'; ob_flush(); flush();
                 } else {
                     echo '==> Cannot Delete File: ' . $realFilePath . '<br />'; ob_flush(); flush();
-                    die('* Please make sure your files are writable. If apache, user:group should be www-data:www-data.');
+                    if (!is_null($phpUser) && $osType == 'linux') {
+                        die('* Please make sure your files are writable. File owner permissions user:group should be ' . $phpUser.':'.$phpUser);
+                    } else if ($osType == 'linux') {
+                        die('* Please make sure your files are writable. Ownership user:group should be www-data:www-data.');
+                    } else {
+                        die('* Please make sure your files are writable.');
+                    }
                 }
             }
         }
@@ -174,7 +213,13 @@ if (isset($_POST['cryptoglance_version']) &&
                 } else {
                     $failedFolders[] = $realFilePath;
                     echo '==> Cannot Delete Folder: ' . $realFilePath . '<br />'; ob_flush(); flush();
-                    die('* Please make sure your directories are writable. If apache, user:group should be www-data:www-data.');
+                    if (!is_null($phpUser) && $osType == 'linux') {
+                        die('* Please make sure your directories are writable. File owner permissions user:group should be ' . $phpUser.':'.$phpUser);
+                    } else if ($osType == 'linux') {
+                        die('* Please make sure your directories are writable. Ownership user:group should be www-data:www-data.');
+                    } else {
+                        die('* Please make sure your directories are writable.');
+                    }
                 }
             }
         }
@@ -185,7 +230,13 @@ if (isset($_POST['cryptoglance_version']) &&
                 echo '==> Deleted Folder: ' . $folderPath . '<br />'; ob_flush(); flush();
             } else {
                 echo '==> Cannot Delete Folder: ' . $folderPath . '<br />'; ob_flush(); flush();
-                die('* Please make sure your directories are writable. If apache, user:group should be www-data:www-data.');
+                if (!is_null($phpUser) && $osType == 'linux') {
+                    die('* Please make sure your directories are writable. File owner permissions user:group should be ' . $phpUser.':'.$phpUser);
+                } else if ($osType == 'linux') {
+                    die('* Please make sure your directories are writable. Ownership user:group should be www-data:www-data.');
+                } else {
+                    die('* Please make sure your directories are writable.');
+                }
             }
         }
         echo '==> Done deleting old files...<br />'; ob_flush(); flush(); sleep(1);
@@ -227,6 +278,9 @@ if (isset($_POST['cryptoglance_version']) &&
         echo '----------<br />';
         echo '==> Cleaning up... <br />'; ob_flush(); flush(); sleep(1);
         echo '==> Deleting update files:<br />'; ob_flush(); flush(); sleep(1);
+        if (unlink($currentDir . $updateDir.'.zip')) {
+            echo '==> Deleted File: ' . $currentDir . $updateDir.'.zip' . '<br />'; ob_flush(); flush();
+        }
         $it = new RecursiveDirectoryIterator($currentDir . 'update', RecursiveDirectoryIterator::SKIP_DOTS);
         $files = new RecursiveIteratorIterator($it, RecursiveIteratorIterator::CHILD_FIRST);
         // first delete files
@@ -273,15 +327,25 @@ if (isset($_POST['cryptoglance_version']) &&
         } else {
             echo '==> Cannot Delete Folder: ' . $currentDir.'update' . '<br />'; ob_flush(); flush();
         }
-        echo '==> Done cleaning up...<br />'; ob_flush(); flush(); sleep(1);
+        echo '==> Done cleaning up...<br />'; ob_flush(); flush(); sleep(2);
 
         // FINISHED -----------------------
 
         echo '----------<br />';
-        echo '** cryptoGlance was updated successfully! **<br />You can navigate back to the Dashboard now.';
+        echo '** cryptoGlance was updated successfully! **';
         echo '<div id="done"></div>'; // this is purely for JS to stop scrolling down
         ob_flush();
         flush();
+
+        if (isset($_SERVER['HTTP_COOKIE'])) {
+            $cookies = explode(';', $_SERVER['HTTP_COOKIE']);
+            foreach($cookies as $cookie) {
+                $parts = explode('=', $cookie);
+                $name = trim($parts[0]);
+                setcookie($name, '', time()-1000);
+                setcookie($name, '', time()-1000, '/');
+            }
+        }
     } else {
         echo "==> ERROR: Updates not enabled!";
         ob_flush();
@@ -289,6 +353,36 @@ if (isset($_POST['cryptoglance_version']) &&
     }
     ob_end_flush();
     echo "</pre>
+    </body>
+</html>";
+    exit;
+} else if ($_GET['view'] == 'update') {
+    echo "<html>
+    <head>
+        <script type=\"text/javascript\" src=\"js/packages/jquery-1.10.2.min.js\"></script>
+        <script type=\"text/javascript\">
+            $(document).ready(function() {
+                setInterval(function() {
+                    $('pre', 'body').append(' .');
+                }, 2500);
+            });
+        </script>
+        <style type=\"text/css\">
+            body {
+              background: #030;
+              color: #090;
+              display: block;
+              font-size: 15px;
+              line-height: 26px;
+              padding: 0 10px;
+              text-shadow: 0px 0px 1px rgba(0, 255, 0, 0.3);
+              text-align: left;
+              text-transform: none;
+            }
+        </style>
+    </head>
+    <body>
+        <pre style=\"font-family: Menlo,Monaco,Consolas,Courier New,monospace;margin:0;white-space: pre-wrap;\">==> Loading, please wait . . .</pre>
     </body>
 </html>";
     exit;
@@ -314,7 +408,7 @@ require_once("includes/header.php");
                     <p>It is advised that you backup your <span class="blue">/<?php echo DATA_FOLDER; ?></span> folder prior to running the update process.</p>
                     <button type="button" class="btn btn-primary" id="btn-update-process" onClick="submit()"><i class="icon icon-restart"></i> Start Update Process</button>
                     <input type="hidden" name="cryptoglance_version" value="<?php echo $_COOKIE['cryptoglance_version'] ?>" />
-                    <iframe name="cg_update" src="#" style="display:none;overflow:hidden;" scrolling="auto"></iframe>
+                    <iframe name="cg_update" src="update.php?view=update" style="display:none;overflow:hidden;" scrolling="auto"></iframe>
                   </form>
                 </div>
               </div>